refactor(cloudformation): move opt-in/opt-out helpers from cfn.go to opt.go
